Vulkan graphics device created, vulkan instance created

+++ DopeEngine/src/Engine/Math/Math.cpp
@@ -3,40 +3,40 @@
 
 namespace DopeEngine
 {
-	 float Math::square_root(float value)
+	float Math::square_root(float value)
 	{
 		return sqrt(value);
 	}
-	 float Math::cos(float rads)
+	float Math::cos(float rads)
 	{
-		return cos(rads);
+		return cosf(rads);
 	}
 	float Math::sin(float rads)
 	{
-		return sin(rads);
+		return sinf(rads);
 	}
-	 float Math::tan(float rads)
+	float Math::tan(float rads)
 	{
-		return tan(rads);
+		return tanf(rads);
 	}
-	 float Math::linear_interpolate(float from, float to, float alpha)
+	float Math::linear_interpolate(float from, float to, float alpha)
 	{
 		const float direction = to - from;
 		return from + direction * alpha;
 	}
-	 float Math::absolute(float value)
+	float Math::absolute(float value)
 	{
 		return (value > 0) ? value : -value;
 	}
-	 float Math::clamp(float value, float min, float max)
+	float Math::clamp(float value, float min, float max)
 	{
 		return value > max ? max : value < min ? min : value;
 	}
-	 float Math::min(float a, float b)
+	float Math::min(float a, float b)
 	{
 		return a > b ? b : a;
 	}
-	 float Math::max(float a, float b)
+	float Math::max(float a, float b)
 	{
 		return a > b ? a : b;
 	}
